Adjust out_type for E2Bottleneck
In e2cnn_exp code, `out_fiber` is actually really important only for the very last layer of the network when we need to output invariant features. However, they also use it to set self.out_type of the block which is used for checking field type in SequentialModule. Because of that, at the end of the block __init__ we need to add `self.out_type = exp_out_fiber`. Since it's only necessary for the last block in the outer layer, we don't pass anymore `out_fiber` to `first_block` and `next_block` inside `_make_layer()`.