Fix mismatch between output of first block and input of second block for fixparams=True
The problem occurs due to rounding error. When fixparams=True, we multiply the number of channels by sqrt(N) and then round it with int(). For example, the output of the first block is calculated in the following way: int[64/8*sqrt(8)]*4=88, while the input to the second block is calculated as int[64*4/8*sqrt(8)]=90 (given inplanes=64, N=8, expansion=4).